requirements.txt
# Deep Learning Framework
torch>=2.5.0
torchvision>=0.20.0
torchaudio>=2.5.0

# Reinforcement Learning Environment
gymnasium>=0.29.0
flappy-bird-gymnasium

# Configuration Management
pyyaml>=6.0

# Numerical Computing
numpy>=1.24.0

# Optional: For visualization and logging
matplotlib>=3.7.0
tensorboard>=2.15.0